requirements.txt
mido>=1.3.0
python-rtmidi>=1.4.11
PyYAML>=6.0
rich>=13.0.0
textual>=0.26.0
